requirements.txt
torch
torchvision
flask
flask-cors
Pillow
efficientnet-pytorch
